pyproject.toml
[project]
name = "workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.2.0",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.4",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.5.3",
    "pymupdf>=1.26.3",
]
